Move `parse_cfg_select` to `rustc_builtin_macros`
